requirements.txt
astroid==2.4.2
autopep8==1.5.6
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
fastcore==1.3.19
flake8==3.9.2
idna==2.10
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617121002983/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort==5.5.2
jedi==0.17.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213356021/work
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.4.2
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
pandas==1.2.4
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.2.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pylint==2.6.0
pyparsing==2.4.7
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz==2021.1
pywin32==227
pyzmq==20.0.0
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
six @ file:///C:/ci/six_1605187374963/work
toml==0.10.1
tornado @ file:///C:/ci/tornado_1606942392901/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
wincertstore==0.2
wrapt==1.12.1
yapf==0.31.0
